backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
marshmallow==3.20.1
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0
